stboot-ci: initial commit
Add scripts to be used by an upcoming CI pipeline.

The output will be:
* stboot image with network support
* stboot image with debug and network support
* stboot image with local os-pkg (ubuntu-20)
* stboot image with debug and local os-pkg (ubuntu-20)
* os-pkg (ubuntu-20)

Note that the debug variants are still missing.

There is also a script "dump_stboot_img" that simply lists all the
files of the image.

The idea is to have the "make keygen-sign" output (cert/keys) reside
somewhere statically, and reuse it for all os-pkgs. Obviously this
need to change later.

Signed-off-by: Björn Töpel <bjorn@mullvad.net>